requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy==1.4.39
psycopg2==2.9.3
debugpy==1.6.2
fastapi-jwt-auth==0.5.0
alembic==1.8.1
python-dotenv==0.20.0
black==22.6.0
colorama==0.4.5
numpy==1.23.3
python-dateutil==2.8.2
unidecode==1.3.6
dependency-injector==4.40.0
pandas==1.5.2
python-multipart==0.0.5
openpyxl==3.0.10